refactor: オーケストレーターコンポーネントをcomponents/直下に移動
components/を開いた時にオーケストレーター（Screen, Header）が
一目でわかるよう、サブディレクトリから直下に移動。
サブディレクトリには部品コンポーネントのみ残す構成に変更。
